+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/UI/ChaosHUD.lua
@@ -127,7 +127,7 @@ function ChaosHUD:prerender()
         if fgWidth > 0 then
             local c = uiCfg.progress_bar_rgb
             local sm = ChaosConfig.streamer_mode
-            if ChaosConfig.use_voting_progress_bar_color and sm and sm.streamer_mode_enabled == true and sm.voting_enabled == true and ChaosEffectsManager.lastVotingActive == 1 then
+            if ChaosConfig.use_voting_progress_bar_color and sm and sm.streamer_mode_enabled == true and sm.voting_enabled == true and ChaosEffectsManager.voteStartedThisInterval then
                 c = uiCfg.progress_bar_voting_rgb
             end
             self:drawRect(0, barY, fgWidth, barHeight, uiCfg.progress_bar_opacity, c.r, c.g, c.b)